+++ src/tool.rs
@@ -1,0 +1,4 @@
+//! This module contains interactive tools, like
+//! color palette generators.
+
+pub mod color;